+++ src/main/java/capstone/domain/Order.java
@@ -22,7 +22,7 @@ public class Order {
     private String symbol;           // 종목코드
     private String name;             // 종목명
     private String type;             // BUY or SELL
-    private Integer quantity;        // 주문수량
+    private Long quantity;        // 주문수량
     private Double price;            // 체결가
     private LocalDateTime createdAt; // 주문시간
